refactor: remove map subcomponents to components folder, create map utils folder
