--- src/components/Map/components/index.ts
@@ -1,1 +1,0 @@
-export { default as ClickablePopup } from './ClickablePopup';
